+++ src/epics_containers_cli/shell.py
@@ -3,14 +3,11 @@ functions for executing commands and querying environment in the linux shell
 """
 
 import os
-import re
 import subprocess
-from pathlib import Path
-from typing import Tuple, Union
+from typing import Union
 
 import typer
 
-from .globals import Architecture
 from .logging import log
 
 EC_GIT_ORG = os.environ.get("EC_GIT_ORG", "")
@@ -45,7 +42,7 @@ def run_command(command: str, interactive=True, error_OK=False) -> Union[str, bo
     output = "" if interactive else p_result.stdout.decode() + p_result.stderr.decode()
 
     if p_result.returncode != 0 and not error_OK:
-        log.error(f"Command Failed:\n{output}")
+        log.error(f"Command Failed:\n{command}\n{output}\n")
         raise typer.Exit(1)
 
     if interactive:
@@ -54,71 +51,3 @@ def run_command(command: str, interactive=True, error_OK=False) -> Union[str, bo
         result = p_result.stdout.decode() + p_result.stderr.decode()
     log.debug(f"returning: {result}")
     return result
-
-
-def get_image_name(
-    repo: str, arch: Architecture = Architecture.linux, target: str = "developer"
-) -> str:
-    registry = repo2registry(repo).lower().removesuffix(".git")
-
-    image = f"{registry}-{arch}-{target}"
-    log.info("repo = %s image  = %s", repo, image)
-    return image
-
-
-def get_git_name(folder: Path = Path(".")) -> Tuple[str, Path]:
-    """
-    work out the git repo name and top level folder for a local clone
-    """
-    os.chdir(folder)
-    path = str(run_command("git rev-parse --show-toplevel", interactive=False))
-    git_root = Path(path.strip())
-
-    remotes = str(run_command("git remote -v", interactive=False))
-    log.debug(f"remotes = {remotes}")
-
-    matches = re.findall(r"((?:(?:git@)|(?:http[s]+:\/\/)).*) (?:.fetch.)", remotes)
-
-    if len(matches) > 0:
-        repo_name = str(matches[0])
-    else:
-        log.error(f"folder {folder.absolute()} cannot parse repo name {remotes}")
-        raise typer.Exit(1)
-
-    log.debug(f"repo_name = {repo_name}, git_root = {git_root}")
-    return repo_name, git_root
-
-
-# work out what the registry name is for a given repo remote e.g.
-def repo2registry(repo_name: str) -> str:
-    """convert a repo name to the related a container registry name"""
-
-    log.debug("extracting fields from repo name %s", repo_name)
-
-    match_git = re.match(r"git@([^:]*):(.*)\/(.*)(?:.git)", repo_name)
-    match_http = re.match(r"https:\/\/([^\/]*)\/([^\/]*)\/([^\/]*)", repo_name)
-    for match in [match_git, match_http]:
-        if match is not None:
-            source_reg, org, repo = match.groups()
-            break
-    else:
-        log.error(f"repo {repo_name} is not a valid git remote")
-        raise typer.Exit(1)
-
-    log.debug("source_reg = %s org = %s repo = %s", source_reg, org, repo)
-
-    if not EC_REGISTRY_MAPPING:
-        log.error("environment variable EC_REGISTRY_MAPPING not set")
-        raise typer.Exit(1)
-
-    for mapping in EC_REGISTRY_MAPPING.split():
-        if mapping.split("=")[0] == source_reg:
-            registry = mapping.split("=")[1]
-            registry = f"{registry}/{org}/{repo}"
-            break
-    else:
-        log.error(f"repo {repo_name} does not match any registry mapping")
-        log.error("please update the environment variable EC_REGISTRY_MAPPING")
-        raise typer.Exit(1)
-
-    return registry
